refactor(bijou): split markdown.ts into markdown-parse and markdown-render
Extract internal implementation from the 468-line markdown.ts into two
focused sibling modules:

- markdown-parse.ts: block-level parser (parseBlocks, isBlockStart),
  inline parser (parseInline with styled/plain/accessible variants),
  word wrapping (wordWrap), and the BlockType union type
- markdown-render.ts: block renderer (renderBlocks) with mode-specific
  output for headings, paragraphs, lists, code blocks, blockquotes, HRs

markdown.ts remains the public facade exporting MarkdownOptions and
the markdown() entry point.

Zero test modifications. Zero index.ts modifications.
